+++ src2/console/__init__.py
@@ -1,0 +1,1 @@
+from src2.console.__main__ import Main
